style(gpu_explorations): rename the directory and related import dep
