+++ mopro-core/src/middleware/mod.rs
@@ -1,4 +1,4 @@
 pub mod circom;
 
 #[cfg(feature = "gpu-benchmarks")]
-pub mod gpu_exploration;
+pub mod gpu_explorations;
